requirements.txt
asgiref==3.6.0
Django==4.1.5
python-dateutil==2.8.2
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
google-api-python-client==2.123.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0